refactor: moved file utils to composable mixin
